Struct pharos::ObserveConfig

pub struct ObserveConfig<Event>
where
    Event: Clone + 'static + Send,
{ /* private fields */ }

Configuration for your event stream.

Pass to Observable::observe when subscribing. This let’s you choose the type of channel and let’s you set a filter to ignore certain events.

use pharos::*;

// We choose event type usize for simplicity. You choose whatever type you want here,
// see the bounds on the Event type parameter throughout this library.
//
let mut pharos = Pharos::<usize>::default();

// Use defaults, unbounded channel and no filter.
//
pharos.observe( ObserveConfig::default() );

// Use bounded channel and defaults for other options.
//
pharos.observe( Channel::Bounded(5).into() );

// Use a filter and defaults for other options.
// Will only receive events if they are bigger than three.
//
pharos.observe( Filter::Pointer( |evt| *evt > 3 ).into() );

// Set both channel and filter. Note you can only set one filter per observable.
//
let opts = ObserveConfig::default()

   .channel( Channel::Bounded( 5 ) )
   .filter ( |evt| *evt > 3        )
;

pharos.observe( opts );

Implementations

impl<Event> ObserveConfig<Event>
where Event: Clone + 'static + Send,

pub fn channel(self, channel: Channel) -> Self

Choose which channel implementation to use for your event stream.

pub fn filter(self, filter: fn(_: &Event) -> bool) -> Self

Filter your event stream with a predicate that is a fn pointer. You can only set one filter per observable.

pub fn filter_boxed( self, filter: impl FnMut(&Event) -> bool + Send + 'static, ) -> Self

Filter your event stream with a predicate that is a closure that captures environment. It is preferred to use filter if you can as this will box the closure. You can only set one filter per observable.